# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "operationId": "AlertsController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Alerts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "instrumentKey": {
                            "type": "string"
                          },
                          "condition": {
                            "type": "string",
                            "enum": [
                              "PRICE_ABOVE",
                              "PRICE_BELOW",
                              "CHANGE_PERCENT_ABOVE",
                              "CHANGE_PERCENT_BELOW",
                              "VOLUME_ABOVE",
                              "OI_CHANGE_ABOVE",
                              "IV_ABOVE",
                              "IV_BELOW",
                              "PCR_ABOVE",
                              "PCR_BELOW"
                            ]
                          },
                          "threshold": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "ACTIVE",
                              "TRIGGERED",
                              "DISABLED"
                            ]
                          },
                          "repeat": {
                            "type": "boolean"
                          },
                          "note": {
                            "nullable": true,
                            "type": "string"
                          },
                          "triggeredAt": {
                            "nullable": true,
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "instrumentKey",
                          "condition",
                          "threshold",
                          "status",
                          "repeat",
                          "note",
                          "triggeredAt",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "summary": "List alerts (newest first)",
        "tags": [
          "alerts"
        ]
      },
      "post": {
        "description": "Conditions: PRICE_ABOVE, PRICE_BELOW, CHANGE_PERCENT_ABOVE, CHANGE_PERCENT_BELOW, VOLUME_ABOVE, OI_CHANGE_ABOVE, IV_ABOVE, IV_BELOW, PCR_ABOVE, PCR_BELOW. PCR conditions are evaluated once a minute on the assembled chain; all others on every live update of the key.",
        "operationId": "AlertsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "instrumentKey": {
                    "type": "string",
                    "maxLength": 96
                  },
                  "condition": {
                    "type": "string",
                    "enum": [
                      "PRICE_ABOVE",
                      "PRICE_BELOW",
                      "CHANGE_PERCENT_ABOVE",
                      "CHANGE_PERCENT_BELOW",
                      "VOLUME_ABOVE",
                      "OI_CHANGE_ABOVE",
                      "IV_ABOVE",
                      "IV_BELOW",
                      "PCR_ABOVE",
                      "PCR_BELOW"
                    ]
                  },
                  "threshold": {
                    "type": "number"
                  },
                  "repeat": {
                    "default": false,
                    "type": "boolean"
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 160
                  }
                },
                "required": [
                  "instrumentKey",
                  "condition",
                  "threshold",
                  "repeat"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "instrumentKey": {
                          "type": "string"
                        },
                        "condition": {
                          "type": "string",
                          "enum": [
                            "PRICE_ABOVE",
                            "PRICE_BELOW",
                            "CHANGE_PERCENT_ABOVE",
                            "CHANGE_PERCENT_BELOW",
                            "VOLUME_ABOVE",
                            "OI_CHANGE_ABOVE",
                            "IV_ABOVE",
                            "IV_BELOW",
                            "PCR_ABOVE",
                            "PCR_BELOW"
                          ]
                        },
                        "threshold": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "TRIGGERED",
                            "DISABLED"
                          ]
                        },
                        "repeat": {
                          "type": "boolean"
                        },
                        "note": {
                          "nullable": true,
                          "type": "string"
                        },
                        "triggeredAt": {
                          "nullable": true,
                          "type": "number"
                        },
                        "createdAt": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "id",
                        "instrumentKey",
                        "condition",
                        "threshold",
                        "status",
                        "repeat",
                        "note",
                        "triggeredAt",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`VALIDATION_FAILED`, per-field issues in `details.issues`)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "summary": "Create an alert (max 50 active per session)",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/v1/alerts/{id}": {
      "patch": {
        "operationId": "AlertsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "threshold": {
                    "type": "number"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "DISABLED"
                    ]
                  },
                  "repeat": {
                    "type": "boolean"
                  },
                  "note": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 160
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "instrumentKey": {
                          "type": "string"
                        },
                        "condition": {
                          "type": "string",
                          "enum": [
                            "PRICE_ABOVE",
                            "PRICE_BELOW",
                            "CHANGE_PERCENT_ABOVE",
                            "CHANGE_PERCENT_BELOW",
                            "VOLUME_ABOVE",
                            "OI_CHANGE_ABOVE",
                            "IV_ABOVE",
                            "IV_BELOW",
                            "PCR_ABOVE",
                            "PCR_BELOW"
                          ]
                        },
                        "threshold": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "TRIGGERED",
                            "DISABLED"
                          ]
                        },
                        "repeat": {
                          "type": "boolean"
                        },
                        "note": {
                          "nullable": true,
                          "type": "string"
                        },
                        "triggeredAt": {
                          "nullable": true,
                          "type": "number"
                        },
                        "createdAt": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "id",
                        "instrumentKey",
                        "condition",
                        "threshold",
                        "status",
                        "repeat",
                        "note",
                        "triggeredAt",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`VALIDATION_FAILED`, per-field issues in `details.issues`)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Unknown instrument, expiry or resource",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "summary": "Update threshold / status / repeat / note",
        "tags": [
          "alerts"
        ]
      },
      "delete": {
        "operationId": "AlertsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "deleted": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        }
                      },
                      "required": [
                        "deleted"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Unknown instrument, expiry or resource",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "summary": "Delete an alert",
        "tags": [
          "alerts"
        ]
      }
    },
...
  },
  "info": {
    "title": "TradeOS API",
    "description": "Indian market option-chain terminal — REST contract v1.\n\nEvery response is JSON. Errors share one envelope: `{ \"error\": { \"code\", \"message\", \"details?\" } }`.\n\nRealtime market data does not use REST: see the WebSocket protocol in `docs/websocket-protocol.md` (`/ws/market`).\n\nInstrument keys are TradeOS-owned (`NSE:INDEX:NIFTY50`, `NSE:OPT:NIFTY50:2026-09-15:24000:CE`); provider symbols never appear in this API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "operator-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Operator-Key"
      }
    },
...
  }
}